pubsub_to_bq/requirements.txt
google-cloud-bigquery==3.3.6
